Lift over xAtlas PL fields (genotype likelihoods)
- in case of overlapping variant records, blank them out (FieldCombinationMethod::MISSING)
- abbreviate format field vector outputs that are all missing values
- move format field helpers into genotyper_utils.h to tidy up genotyper.cc
