+++ test/data/gvcf_test_cases/censor_rnc_format_fields.yml
@@ -97,4 +97,4 @@ truth_output_vcf:
       ##FORMAT=<ID=RNC,Number=G,Type=Character,Description="Reason for No Call in GT: . = n/a, M = Missing data, P = Partial data, D = insufficient Depth of coverage, - = unrepresentable overlapping deletion, L = Lost/unrepresentable allele (other than deletion), U = multiple Unphased variants present, O = multiple Overlapping variants present">
       ##contig=<ID=A,length=1000000>
       #CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO	FORMAT	A	B	C	D
-      A	907970	.	AC	AA,A	757	.	.	GT:GQ:DP:AD:SB:RNC	0/0:99:76:21,0,0:.,.,.,.:..	0/1:99:45:17,26,0:3,14,7,19:..	./.:.:.:.,.,.:.,.,.,.:PP	./1:.:27:.,24,3:.,.,.,.:O.
+      A	907970	.	AC	AA,A	757	.	.	GT:GQ:DP:AD:SB:RNC	0/0:99:76:21,0,0:.,.,.,.:..	0/1:99:45:17,26,0:3,14,7,19:..	./.:.:.:.:.,.,.,.:PP	./1:.:27:.,24,3:.,.,.,.:O.
